Work toward resolution of #2281

+++ fastai/callback.py
@@ -287,7 +287,7 @@ class CallbackHandler():
 
     def on_backward_begin(self, loss:Tensor)->Tuple[Any,Any]:
         "Handle gradient calculation on `loss`."
-        self.smoothener.add_value(loss.detach().cpu())
+        self.smoothener.add_value(loss.float().detach().cpu())
         self.state_dict['last_loss'], self.state_dict['smooth_loss'] = loss, self.smoothener.smooth
         self('backward_begin', call_mets=False)
         return self.state_dict['last_loss'], self.state_dict['skip_bwd']
